Establishes system configuration framework
This creates a MySql repository that will store each property of a class
is a database row for a given class name. The repository inherits the
IConfigurationRepository and is used by the SystemConfigurationService
to get and save the SystemConfigurationEntry class.

Also included in this commit are several database model changes to work
with his new method of configuration storage and removes the
PickupSchedule from the StationCollectorRow.